app/requirements.txt
django==4.0.2

python-decouple==3.6

psycopg2==2.9.3
redis==4.1.2
django-redis==5.2.0
celery==5.2.3

djangorestframework==3.13.1
drf-spectacular[sidecar]==0.21.2

flake8==4.0.1
flake8-bugbear==22.1.11
pep8-naming==0.12.1


pytest==6.2.5
pytest-django==4.5.2
mixer==7.2.0
